writing/pyproject.toml
[project]
name = "writing"
version = "0.1.0"
description = "Language Learning Assistant - Generate and translate simple sentences"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "streamlit>=1.30.0",
    "openai>=1.10.0",
    "pillow>=10.0.0",
]
